Saturn rings do still not work (disabled the two sided mode, at least it works on one side now)

+++ OpenGL_SceneGraph_Stussman/Code/Aufgabe4/include/clock.h
@@ -33,7 +33,7 @@ protected:
 
 private:
   double mTime;
-  long mOldNow;
+  int mOldNow;
   float mDilatation;
   std::vector<ClockListener*> mListeners;
